requirements.txt
alembic==1.17.0
annotated-types==0.7.0
anyio==4.11.0
click==8.3.0
fastapi==0.119.1
flake8==7.3.0
greenlet==3.2.4
h11==0.16.0
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
mccabe==0.7.0
mypy==1.18.2
mypy_extensions==1.1.0
pathspec==0.12.1
psycopg==3.2.11
pycodestyle==2.14.0
pydantic==2.12.3
pydantic_core==2.41.4
pyflakes==3.4.0
python-dotenv==1.1.1
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
